requirements.txt
anyio==4.8.0
beautifulsoup4==4.13.3
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
duckduckgo_search==7.4.3
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
mdurl==0.1.2
notion-client==2.3.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
smolagents==1.9.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
